requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
joblib
pymongo
python-dotenv
requests